refactored workers and associated files to single copy for each
