--- connection.js
@@ -1,6 +1,0 @@
-const { Client } = require('pg');
-const client = new Client();
-
-client.connect();
-
-module.exports = client;
