+++ package.json
@@ -3,7 +3,9 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "nodemon ./bin/www"
+    "start": "nodemon ./bin/www",
+    "start-ts": "PGDATABASE=chronos_ts & node workers/worker ts",
+    "start-pl": "PGDATABASE=chronos_pl & node workers/worker pl"
   },
   "dependencies": {
     "amqplib": "^0.5.2",
