+++ workers/writeToDB.js
@@ -1,13 +1,5 @@
-const { Client } = require('pg');
-const client = new Client({
-  user: 'user',
-  host: 'localhost',
-  database: 'chronos_pl',
-  port: '5432',
-});
-client.connect();
-
-const INSERT = require('./pipelinedb_queries');
+const client = require('./client');
+const INSERT = require('./queries');
 
 const writeToDB = (msg) => {
   const json = JSON.parse(msg.content);
@@ -16,23 +8,23 @@ const writeToDB = (msg) => {
   let text;
   let values;
 
-  if (eType === 'clicks'){
-    let { target_node, buttons, x, y } = json;
-    text = INSERT.click;
-    values = [target_node, buttons, x, y, timestamp, metadata];
-  } else if (eType === 'link_clicks') {
+  if (eType === 'link_clicks') {
     let { linkText, targetURL } = json;
     text = INSERT.link_click;
     values = [linkText, targetURL, timestamp, metadata];
+  } else if (eType === 'clicks') {
+    let { target_node, buttons, x, y } = json;
+    text = INSERT.click;
+    values = [target_node, buttons, x, y, timestamp, metadata];
   } else if (eType === 'mouse_moves') {
     let { x, y } = json;
     text = INSERT.mouse_move;
     values = [x, y, timestamp, metadata];
-  } else if (eType === 'key_press') {
+  } else if (eType === 'key_presses') {
     let { key } = json;
     text = INSERT.key_press;
     values = [key, timestamp, metadata];
-  } else if (eType === 'pageview') {
+  } else if (eType === 'pageviews') {
     let { url, title } = json;
     text = INSERT.pageview;
     values = [url, title, timestamp, metadata];
@@ -43,8 +35,8 @@ const writeToDB = (msg) => {
   }
 
   client.query(text, values, (err, res) => {
-    console.log(err ? err.stack : 'Success');
-  })
+    console.log(err ? err.stack : res.rows[0]);
+  });
 }
 
 module.exports = writeToDB;
